Add validation to Change node editor for invalid regex
This change adds input validation to the gui of Change Nodes to prevent
the user from unintentionally entering an invalid regular expression
(in case the ‘use regular expressions’  option is enabled).

The user will be notified (using the RED notification mechanism) on the
specific error code to help resolve the issue.